refactor(gir): Restructure imports and introduce new class and parameter types
- Updated import paths to improve clarity and maintainability, moving to specific files for introspected classes and parameters.
- Introduced new `IntrospectedClassFunction`, `IntrospectedConstructor`, and `IntrospectedFunctionParameter` types to enhance type safety and organization.
- Removed deprecated imports and consolidated related functionality into more specific modules.
- Updated various files to utilize the new types, ensuring consistency across the codebase.
