+++ packages/lib/src/generators/generator.ts
@@ -1,23 +1,21 @@
-import { IntrospectedNamespace } from "../gir/namespace.ts";
-import { IntrospectedClass } from "../gir/introspected-class.ts";
-import { IntrospectedRecord } from "../gir/record.ts";
-import { IntrospectedInterface } from "../gir/interface.ts";
-import { IntrospectedConstant } from "../gir/const.ts";
-import { IntrospectedEnum, IntrospectedError, GirEnumMember } from "../gir/enum.ts";
-import { IntrospectedProperty, IntrospectedField } from "../gir/property.ts";
-import { IntrospectedSignal, IntrospectedSignalType } from "../gir/signal.ts";
-import {
-    IntrospectedFunction,
-    IntrospectedFunctionParameter,
-    IntrospectedConstructor,
-    IntrospectedDirectAllocationConstructor
-} from "../gir/function.ts";
-import { IntrospectedCallback, IntrospectedClassCallback } from "../gir/callback.ts";
-import { IntrospectedClassFunction } from "../gir/function.ts";
-import { IntrospectedStaticClassFunction } from "../gir/function.ts";
-import { IntrospectedVirtualClassFunction } from "../gir/function.ts";
-import { IntrospectedAlias } from "../gir/alias.ts";
-import { TypeExpression } from "../gir.ts";
+import type { IntrospectedNamespace } from "../gir/namespace.ts";
+import type { IntrospectedClass } from "../gir/introspected-class.ts";
+import type { IntrospectedRecord } from "../gir/record.ts";
+import type { IntrospectedInterface } from "../gir/interface.ts";
+import type { IntrospectedConstant } from "../gir/const.ts";
+import type { IntrospectedEnum, IntrospectedError, GirEnumMember } from "../gir/enum.ts";
+import type { IntrospectedProperty, IntrospectedField } from "../gir/property.ts";
+import type { IntrospectedSignal, IntrospectedSignalType } from "../gir/signal.ts";
+import type { IntrospectedFunctionParameter } from "../gir/parameter.ts";
+import type { IntrospectedConstructor } from "../gir/constructor.ts";
+import type { IntrospectedDirectAllocationConstructor } from "../gir/direct-allocation-constructor.ts";
+import type { IntrospectedFunction } from "../gir/function.ts";
+import type { IntrospectedCallback, IntrospectedClassCallback } from "../gir/callback.ts";
+import type { IntrospectedClassFunction } from "../gir/class-function.ts";
+import type { IntrospectedStaticClassFunction } from "../gir/static-function.ts";
+import type { IntrospectedVirtualClassFunction } from "../gir/virtual-function.ts";
+import type { IntrospectedAlias } from "../gir/alias.ts";
+import type { TypeExpression } from "../gir.ts";
 import type { OptionsGeneration } from "../types/options-generation.ts";
 
 export interface GenericDescriptor {
