+++ packages/lib/src/gir/function.ts
@@ -9,40 +9,36 @@ import {
     Generic,
 } from "../gir.ts";
 
-import { IntrospectedBase } from "./introspected-base.ts";
-import { IntrospectedClassMember } from "./introspected-class-member.ts";
 import { IntrospectedNamespaceMember } from "./introspected-namespace-member.ts";
 import type { IntrospectedOptions } from "../types/index.ts";
 
 import type {
     GirFunctionElement,
     GirMethodElement,
-    GirCallableParamElement,
     GirVirtualMethodElement,
-    GirConstructorElement,
     GirModule,
 } from "../index.ts";
 import { GirDirection } from '@gi.ts/parser'
 
 import { IntrospectedNamespace } from "./namespace.ts";
-import { isIntrospectable } from "../utils/index.ts";
-import { getType, isInvalid, sanitizeMemberName, sanitizeIdentifierName, parseDoc, parseMetadata, hasShadow } from "../utils/index.ts";
-import { IntrospectedClass } from "./introspected-class.ts";
+import { isIntrospectable } from "../utils/girs.ts";
+import { getType, hasShadow } from "../utils/gir-parsing.ts";
+import { parseDoc, parseMetadata } from "../utils/gir-parsing.ts";
+import { sanitizeIdentifierName } from "../utils/naming.ts";
+
+
 import { IntrospectedEnum } from "./enum.ts";
-import { IntrospectedSignal } from "./signal.ts";
 import { FormatGenerator } from "../generators/generator.ts";
 import { GirVisitor } from "../visitor.ts";
-import { IntrospectedField } from "./property.ts";
 import { IntrospectedBaseClass } from "./introspected-base-class.ts";
 
 import type { OptionsLoad } from "../types/index.ts";
 
-// Import callback classes from the separate file
-import { IntrospectedCallback, IntrospectedClassCallback } from "./callback.ts";
-
-// ============================================================================
-// BASE FUNCTION CLASSES
-// ============================================================================
+import { IntrospectedCallback } from "./callback.ts";
+import { IntrospectedFunctionParameter } from "./parameter.ts";
+import { IntrospectedClassFunction } from "./class-function.ts";
+import { IntrospectedVirtualClassFunction } from "./virtual-function.ts";
+import { IntrospectedStaticClassFunction } from "./static-function.ts";
 
 export class IntrospectedFunction extends IntrospectedNamespaceMember {
     readonly parameters: IntrospectedFunctionParameter[];
@@ -323,666 +319,6 @@ export class IntrospectedFunction extends IntrospectedNamespaceMember {
     }
 }
 
-// ============================================================================
-// CONSTRUCTOR CLASSES
-// ============================================================================
 
-export class IntrospectedDirectAllocationConstructor extends IntrospectedClassMember {
-    parameters: IntrospectedFunctionParameter[];
-
-    constructor(parameters: IntrospectedFunctionParameter[], parent: IntrospectedBaseClass | null) {
-        super("new", parent, { isPrivate: false, isIntrospectable: true });
-
-        this.parameters = parameters.map(parameter => parameter.copy({ parent: this }));
-    }
-
-    static fromFields(fields: IntrospectedField[], parent: IntrospectedBaseClass) {
-        const building = new IntrospectedDirectAllocationConstructor([], parent);
-
-        building.parameters = fields
-            .filter(field => {
-                return !field.isStatic && !field.isPrivate && !field.type.isPointer;
-            })
-            .map(
-                field =>
-                    new IntrospectedFunctionParameter({
-                        parent: building,
-                        type: field.type,
-                        name: field.name,
-                        direction: GirDirection.In
-                    })
-            );
-
-        return building;
-    }
-
-    asString<T extends FormatGenerator<unknown>>(generator: T): ReturnType<T["generateDirectAllocationConstructor"]> {
-        return generator.generateDirectAllocationConstructor(this) as ReturnType<
-            T["generateDirectAllocationConstructor"]
-        >;
-    }
-
-    copy(options?: {
-        parent?: IntrospectedBaseClass | undefined;
-        parameters?: IntrospectedFunctionParameter[];
-    }): IntrospectedDirectAllocationConstructor {
-        const copy = new IntrospectedDirectAllocationConstructor(
-            options?.parameters ?? this.parameters,
-            options?.parent ?? this.parent
-        );
-
-        copy._copyBaseProperties(this);
-
-        return copy;
-    }
-
-    accept(visitor: GirVisitor): IntrospectedDirectAllocationConstructor {
-        const node = this.copy({
-            parameters: this.parameters.map(parameters => {
-                return parameters.accept(visitor);
-            })
-        });
-
-        return visitor.visitDirectAllocationConstructor?.(node) ?? node;
-    }
-}
-
-export class IntrospectedConstructor extends IntrospectedClassMember {
-    readonly parameters: IntrospectedFunctionParameter[] = [];
-    readonly return_type: TypeExpression = UnknownType;
-
-    generics: Generic[] = [];
-
-    constructor({
-        name,
-        parameters = [],
-        return_type,
-        parent,
-        ...args
-    }: IntrospectedOptions<{
-        name: string;
-        parent: IntrospectedBaseClass | null;
-        parameters?: IntrospectedFunctionParameter[];
-        return_type: TypeExpression;
-    }>) {
-        super(name, parent, { ...args });
-        this.return_type = return_type;
-        this.parameters = parameters.map(p => p.copy({ parent: this }));
-    }
-
-    copy({
-        parent,
-        parameters,
-        return_type
-    }: {
-        parent?: IntrospectedBaseClass;
-        parameters?: IntrospectedFunctionParameter[];
-        return_type?: TypeExpression;
-    } = {}): IntrospectedConstructor {
-        const constr = new IntrospectedConstructor({
-            name: this.name,
-            parent: parent ?? this.parent ?? null,
-            return_type: return_type ?? this.return_type,
-            parameters: parameters ?? this.parameters
-        })._copyBaseProperties(this);
-
-        constr.generics = [...this.generics];
-
-        return constr;
-    }
-
-    static fromXML(
-        m: GirConstructorElement,
-        parent: IntrospectedBaseClass,
-        options: OptionsLoad
-    ): IntrospectedConstructor {
-        return IntrospectedClassFunction.fromXML(m as GirFunctionElement, parent, options).asConstructor();
-    }
-
-    accept(visitor: GirVisitor): IntrospectedConstructor {
-        const node = this.copy({
-            parameters: this.parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            return_type: visitor.visitType?.(this.return_type)
-        });
-
-        return visitor.visitConstructor?.(node) ?? node;
-    }
-
-    return() {
-        return this.return_type;
-    }
-
-    asString<T extends FormatGenerator<unknown>>(generator: T): ReturnType<T["generateConstructor"]> {
-        return generator.generateConstructor(this) as ReturnType<T["generateConstructor"]>;
-    }
-}
-
-// ============================================================================
-// FUNCTION PARAMETER
-// ============================================================================
-
-export class IntrospectedFunctionParameter extends IntrospectedBase<
-    | IntrospectedClassFunction
-    | IntrospectedFunction
-    | IntrospectedSignal
-    | IntrospectedConstructor
-    | IntrospectedDirectAllocationConstructor
-    | null
-> {
-    readonly type: TypeExpression;
-    readonly direction: GirDirection;
-    readonly isVarArgs: boolean = false;
-    readonly isOptional: boolean = false;
-    readonly isNullable: boolean = false;
-
-    constructor({
-        name,
-        direction,
-        type,
-        parent,
-        doc,
-        isVarArgs = false,
-        isOptional = false,
-        isNullable = false,
-        ...args
-    }: IntrospectedOptions<{
-        name: string;
-        parent?:
-            | IntrospectedClassFunction
-            | IntrospectedFunction
-            | IntrospectedSignal
-            | IntrospectedConstructor
-            | IntrospectedDirectAllocationConstructor
-            | null;
-        type: TypeExpression;
-        direction: GirDirection;
-        doc?: string | null;
-        isVarArgs?: boolean;
-        isOptional?: boolean;
-        isNullable?: boolean;
-    }>) {
-        super(name, parent ?? null, { ...args });
-
-        this.type = type;
-        this.direction = direction;
-        this.doc = doc;
-        this.isVarArgs = isVarArgs;
-        this.isOptional = isOptional;
-        this.isNullable = isNullable;
-    }
-
-    get namespace() {
-        if (!this.parent) throw new Error(`Failed to get namespace for ${this.name}`);
-
-        return this.parent.namespace;
-    }
-
-    asField() {
-        const { name, parent, isOptional: optional, type } = this;
-
-        if (!(parent?.parent instanceof IntrospectedBaseClass)) {
-            throw new Error("Can't convert parameter of non-class function to field");
-        }
-
-        return new IntrospectedField({ name, parent: parent.parent, optional, type });
-    }
-
-    copy(
-        options: {
-            parent?:
-                | IntrospectedClassFunction
-                | IntrospectedFunction
-                | IntrospectedSignal
-                | IntrospectedConstructor
-                | IntrospectedDirectAllocationConstructor
-                | null;
-            type?: TypeExpression;
-            isOptional?: boolean;
-            isNullable?: boolean;
-        } = {
-            parent: this.parent
-        }
-    ): IntrospectedFunctionParameter {
-        const { type, parent, direction, isVarArgs, isOptional, isNullable, name, doc } = this;
-
-        return new IntrospectedFunctionParameter({
-            parent: options.parent ?? parent,
-            name,
-            direction,
-            isVarArgs,
-            isOptional: options.isOptional ?? isOptional,
-            isNullable: options.isNullable ?? isNullable,
-            type: options.type ?? type,
-            doc
-        })._copyBaseProperties(this);
-    }
-
-    accept(visitor: GirVisitor): IntrospectedFunctionParameter {
-        const node = this.copy({
-            type: visitor.visitType?.(this.type)
-        });
-
-        return visitor.visitParameter?.(node) ?? node;
-    }
-
-    asString<T>(generator: FormatGenerator<T>): T {
-        return generator.generateParameter(this);
-    }
-
-    static fromXML<
-        Parent extends IntrospectedSignal | IntrospectedClassFunction | IntrospectedFunction | IntrospectedConstructor
-    >(
-        parameter: GirCallableParamElement & { $: { name: string } },
-        parent: Parent,
-        options: OptionsLoad
-    ): IntrospectedFunctionParameter {
-        const ns = parent.namespace;
-        let name = sanitizeMemberName(parameter.$.name);
-
-        if (isInvalid(name)) {
-            name = `_${name}`;
-        }
-
-        let isVarArgs = false;
-        let isOptional = false;
-        let isNullable = false;
-
-        let type: TypeExpression;
-        let direction: GirDirection;
-
-        if (
-            !parameter.$.direction ||
-            parameter.$.direction === GirDirection.In ||
-            parameter.$.direction === GirDirection.Inout ||
-            parameter.$.direction === GirDirection.InOut
-        ) {
-            if (name === "...") {
-                isVarArgs = true;
-                name = "args";
-            }
-
-            // Default to "in" direction
-            direction = parameter.$.direction || GirDirection.In;
-
-            const optional = parameter.$.optional === "1";
-            const nullable = parameter.$.nullable === "1";
-
-            if (optional) {
-                isOptional = true;
-            }
-
-            if (nullable) {
-                isNullable = true;
-            }
-
-            type = getType(ns, parameter);
-        } else if (parameter.$.direction === GirDirection.Out) {
-            direction = parameter.$.direction;
-            type = getType(ns, parameter);
-        } else {
-            throw new Error(`Unknown parameter direction: ${parameter.$.direction as string}`);
-        }
-
-        const fp = new IntrospectedFunctionParameter({
-            isVarArgs,
-            type: isVarArgs ? new ArrayType(type) : type,
-            direction,
-            parent: parent ?? undefined,
-            isOptional,
-            isNullable,
-            name,
-            isIntrospectable: isIntrospectable(parameter)
-        });
-
-        if (options.loadDocs) {
-            fp.doc = parseDoc(parameter);
-            fp.metadata = parseMetadata(parameter);
-        }
-
-        return fp;
-    }
-}
-
-// ============================================================================
-// CLASS FUNCTION CLASSES
-// ============================================================================
-
-export class IntrospectedClassFunction<
-    Parent extends IntrospectedBaseClass | IntrospectedEnum = IntrospectedBaseClass | IntrospectedEnum
-> extends IntrospectedBase<Parent> {
-    readonly parameters: IntrospectedFunctionParameter[];
-    protected readonly return_type: TypeExpression;
-    readonly output_parameters: IntrospectedFunctionParameter[];
-    protected _anyify: boolean = false;
-    protected _generify: boolean = false;
-    interfaceParent: IntrospectedBaseClass | IntrospectedEnum | null = null;
-    returnTypeDoc?: string | null;
-
-    generics: Generic[] = [];
-
-    constructor({
-        name,
-        parameters = [],
-        output_parameters = [],
-        return_type = UnknownType,
-        parent,
-        doc,
-        ...args
-    }: IntrospectedOptions<{
-        name: string;
-        parameters?: IntrospectedFunctionParameter[];
-        output_parameters?: IntrospectedFunctionParameter[];
-        return_type?: TypeExpression;
-        parent: Parent;
-        originalParent?: Parent | null;
-        doc?: string | null;
-    }>) {
-        super(name, parent, { ...args });
-
-        this.parameters = parameters.map(p => p.copy({ parent: this }));
-        this.output_parameters = output_parameters.map(p => p.copy({ parent: this }));
-        this.return_type = return_type;
-        this.doc = doc;
-    }
-
-    get namespace() {
-        return this.parent.namespace;
-    }
-
-    asCallback(): IntrospectedClassCallback {
-        const { name, parent, output_parameters, parameters, return_type } = this;
-
-        return new IntrospectedClassCallback({
-            parent,
-            name,
-            output_parameters,
-            parameters,
-            return_type
-        });
-    }
-
-    asConstructor(): IntrospectedConstructor {
-        const { name, parent, parameters } = this;
-
-        if (parent instanceof IntrospectedBaseClass) {
-            // Always force constructors to have the correct return type.
-            return new IntrospectedConstructor({
-                name,
-                parent,
-                parameters,
-                return_type: this.parent.getType(),
-                isIntrospectable: this.isIntrospectable
-            });
-        }
-
-        throw new Error(
-            `Attempted to convert GirClassFunction into GirConstructor from invalid enum parent: ${this.parent.name}`
-        );
-    }
-
-    asStaticClassFunction(parent: IntrospectedClass) {
-        const { name, parameters, return_type, output_parameters } = this;
-
-        return new IntrospectedStaticClassFunction({
-            name,
-            parameters,
-            output_parameters,
-            return_type,
-            parent,
-            isIntrospectable: this.isIntrospectable
-        });
-    }
-
-    copy({
-        parent = this.parent,
-        name,
-        interfaceParent,
-        parameters,
-        outputParameters,
-        returnType
-    }: {
-        parent?: Parent;
-        name?: string;
-        interfaceParent?: IntrospectedBaseClass | IntrospectedEnum;
-        parameters?: IntrospectedFunctionParameter[];
-        outputParameters?: IntrospectedFunctionParameter[];
-        returnType?: TypeExpression;
-    } = {}): IntrospectedClassFunction<Parent> {
-        const fn = new IntrospectedClassFunction<Parent>({
-            name: name ?? this.name,
-            parent,
-            output_parameters: outputParameters ?? this.output_parameters,
-            parameters: parameters ?? this.parameters,
-            return_type: returnType ?? this.return_type
-        });
-
-        fn.generics = [...this.generics];
-        fn.returnTypeDoc = this.returnTypeDoc;
-
-        if (interfaceParent) {
-            fn.interfaceParent = interfaceParent;
-        }
-
-        return fn._copyBaseProperties(this);
-    }
-
-    accept(visitor: GirVisitor): IntrospectedClassFunction<Parent> {
-        const node = this.copy({
-            parameters: this.parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            outputParameters: this.output_parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            returnType: visitor.visitType?.(this.return_type)
-        });
-
-        const fn = visitor.visitClassFunction?.(node);
-
-        return fn ?? node;
-    }
-
-    anyify(): this {
-        this._anyify = true;
-
-        return this;
-    }
-
-    shouldAnyify() {
-        return this._anyify;
-    }
-
-    static fromXML(
-        element: GirFunctionElement | GirMethodElement,
-        parent: IntrospectedBaseClass | IntrospectedEnum,
-        options: OptionsLoad
-    ): IntrospectedClassFunction {
-        const fn = IntrospectedFunction.fromXML(element, parent.namespace, options);
-
-        return fn.asClassFunction(parent);
-    }
-
-    return(): TypeExpression {
-        return this.return_type;
-    }
-
-    asString<T extends FormatGenerator<unknown>>(generator: T): ReturnType<T["generateClassFunction"]> {
-        return generator.generateClassFunction(this) as ReturnType<T["generateClassFunction"]>;
-    }
-}
-
-export class IntrospectedVirtualClassFunction extends IntrospectedClassFunction<IntrospectedBaseClass> {
-    constructor({
-        name,
-        parameters = [],
-        output_parameters = [],
-        return_type = UnknownType,
-        parent,
-        doc,
-        ...args
-    }: IntrospectedOptions<{
-        name: string;
-        parameters: IntrospectedFunctionParameter[];
-        output_parameters?: IntrospectedFunctionParameter[];
-        return_type?: TypeExpression;
-        parent: IntrospectedBaseClass;
-        doc?: string | null;
-    }>) {
-        super({
-            parent,
-            name: name.startsWith("vfunc_") ? name : `vfunc_${name}`,
-            parameters,
-            output_parameters,
-            return_type,
-            doc,
-            ...args
-        });
-    }
-
-    copy({
-        parent = this.parent,
-        interfaceParent,
-        parameters,
-        outputParameters,
-        returnType
-    }: {
-        parent?: IntrospectedBaseClass;
-        interfaceParent?: IntrospectedBaseClass | IntrospectedEnum | undefined;
-        parameters?: IntrospectedFunctionParameter[] | undefined;
-        outputParameters?: IntrospectedFunctionParameter[] | undefined;
-        returnType?: TypeExpression | undefined;
-    }): IntrospectedVirtualClassFunction {
-        const fn = new IntrospectedVirtualClassFunction({
-            name: this.name,
-            parent,
-            output_parameters: outputParameters ?? this.output_parameters,
-            parameters: parameters ?? this.parameters,
-            return_type: returnType ?? this.return_type
-        });
-
-        fn.generics = [...this.generics];
-        fn.returnTypeDoc = this.returnTypeDoc;
-
-        if (interfaceParent) {
-            fn.interfaceParent = interfaceParent;
-        }
-
-        return fn._copyBaseProperties(this);
-    }
-
-    return(): TypeExpression {
-        return this.return_type;
-    }
-
-    accept(visitor: GirVisitor): IntrospectedVirtualClassFunction {
-        const node = this.copy({
-            parameters: this.parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            outputParameters: this.output_parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            returnType: visitor.visitType?.(this.return_type)
-        });
-        return visitor.visitVirtualClassFunction?.(node) ?? node;
-    }
-
-    static fromXML(
-        m: GirVirtualMethodElement,
-        parent: IntrospectedBaseClass,
-        options: OptionsLoad
-    ): IntrospectedVirtualClassFunction {
-        const fn = IntrospectedFunction.fromXML(m, parent.namespace, options);
-
-        return fn.asVirtualClassFunction(parent);
-    }
-
-    asString<T extends FormatGenerator<unknown>>(generator: T): ReturnType<T["generateVirtualClassFunction"]> {
-        return generator.generateVirtualClassFunction(this) as ReturnType<T["generateVirtualClassFunction"]>;
-    }
-}
-
-export class IntrospectedStaticClassFunction extends IntrospectedClassFunction {
-    asString<T extends FormatGenerator<unknown>>(generator: T): ReturnType<T["generateStaticClassFunction"]> {
-        return generator.generateStaticClassFunction(this) as ReturnType<T["generateStaticClassFunction"]>;
-    }
-
-    copy({
-        parent = this.parent,
-        interfaceParent,
-        parameters,
-        outputParameters,
-        returnType
-    }: {
-        parent?: IntrospectedBaseClass | IntrospectedEnum;
-        interfaceParent?: IntrospectedBaseClass | IntrospectedEnum | undefined;
-        parameters?: IntrospectedFunctionParameter[] | undefined;
-        outputParameters?: IntrospectedFunctionParameter[] | undefined;
-        returnType?: TypeExpression | undefined;
-    } = {}): IntrospectedStaticClassFunction {
-        const fn = new IntrospectedStaticClassFunction({
-            name: this.name,
-            parent,
-            output_parameters: outputParameters ?? this.output_parameters,
-            parameters: parameters ?? this.parameters,
-            return_type: returnType ?? this.return_type
-        });
-
-        fn.generics = [...this.generics];
-        fn.returnTypeDoc = this.returnTypeDoc;
-
-        if (interfaceParent) {
-            fn.interfaceParent = interfaceParent;
-        }
-
-        return fn._copyBaseProperties(this);
-    }
-
-    accept(visitor: GirVisitor): IntrospectedStaticClassFunction {
-        const node = this.copy({
-            parent: this.parent,
-            parameters: this.parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            outputParameters: this.output_parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            returnType: visitor.visitType?.(this.return_type)
-        });
-
-        return visitor.visitStaticClassFunction?.(node) ?? node;
-    }
-
-    asClassFunction(parent: IntrospectedBaseClass): IntrospectedClassFunction {
-        const { name, output_parameters, parameters, return_type, doc, isIntrospectable } = this;
-
-        const fn = new IntrospectedClassFunction({
-            parent,
-            name,
-            output_parameters,
-            parameters,
-            return_type,
-            doc,
-            isIntrospectable
-        });
-
-        fn.returnTypeDoc = this.returnTypeDoc;
-        fn.generics = [...this.generics];
-
-        return fn;
-    }
-
-    static fromXML(
-        m: GirFunctionElement,
-        parent: IntrospectedBaseClass | IntrospectedEnum,
-        options: OptionsLoad
-    ): IntrospectedStaticClassFunction {
-        const fn = IntrospectedFunction.fromXML(m, parent.namespace, options);
-
-        return fn.asStaticClassFunction(parent);
-    }
-}
