+++ packages/lib/src/gir/index.ts
@@ -1,10 +1,12 @@
 export * from "./alias.ts";
 export * from "./callback.ts";
+export * from "./class-function.ts";
+export * from "./direct-allocation-constructor.ts";
+export * from "./function.ts";
+export * from "./constructor.ts";
 export * from "./const.ts";
 export * from "./data.ts";
 export * from "./enum.ts";
-export * from "./function.ts";
-export * from "./callback.ts";
 export * from "./generics.ts";
 export * from "./gir-complex-record.ts";
 export * from "./interface.ts";
@@ -14,8 +16,11 @@ export * from "./introspected-class-member.ts";
 export * from "./introspected-class.ts";
 export * from "./introspected-namespace-member.ts";
 export * from "./namespace.ts";
+export * from "./parameter.ts";
 export * from "./promisify.ts";
 export * from "./property.ts";
 export * from "./record.ts";
 export * from "./registry.ts";
 export * from "./signal.ts";
+export * from "./static-function.ts";
+export * from "./virtual-function.ts";
