+++ packages/lib/src/gir/introspected-base-class.ts
@@ -16,9 +16,9 @@ import {
 
 import {
     IntrospectedClassFunction,
-    IntrospectedConstructor,
-    IntrospectedDirectAllocationConstructor
-} from "./function.ts";
+} from "./class-function.ts";
+import { IntrospectedConstructor } from "./constructor.ts";
+import { IntrospectedDirectAllocationConstructor } from "./direct-allocation-constructor.ts";
 import { IntrospectedClassCallback } from "./callback.ts";
 import { IntrospectedProperty, IntrospectedField } from "./property.ts";
 import { IntrospectedNamespace } from "./namespace.ts";
