+++ packages/lib/src/gir/introspected-class.ts
@@ -13,20 +13,20 @@ import {
 
 import {
     IntrospectedClassFunction,
-    IntrospectedVirtualClassFunction,
-    IntrospectedStaticClassFunction,
-    IntrospectedConstructor
-} from "./function.ts";
+    
+} from "./class-function.ts";
+import { IntrospectedVirtualClassFunction } from "./virtual-function.ts";
+import { IntrospectedStaticClassFunction } from "./static-function.ts";
+import { IntrospectedConstructor } from "./constructor.ts";
+
 import { IntrospectedClassCallback } from "./callback.ts";
 import { IntrospectedProperty, IntrospectedField } from "./property.ts";
 import { IntrospectedNamespace } from "./namespace.ts";
-import {
-    sanitizeIdentifierName,
-    parseTypeIdentifier,
-    resolveTypeIdentifier,
-    parseDoc,
-    parseMetadata
-} from "../utils/index.ts";
+import { sanitizeIdentifierName } from "../utils/naming.ts";
+import { parseDoc, parseMetadata } from "../utils/gir-parsing.ts";
+import { resolveTypeIdentifier } from "../utils/type-resolution.ts";
+import { parseTypeIdentifier } from "../utils/types.ts";
+
 import { IntrospectedSignal } from "./signal.ts";
 import { FormatGenerator } from "../generators/generator.ts";
 import { GirVisitor } from "../visitor.ts";
