+++ packages/lib/src/gir/record.ts
@@ -16,20 +16,22 @@ import {
 
 import {
     IntrospectedClassFunction,
-    IntrospectedVirtualClassFunction,
-    IntrospectedStaticClassFunction,
-    IntrospectedConstructor,
-} from "./function.ts";
+} from "./class-function.ts";
+import { IntrospectedConstructor } from "./constructor.ts";
+import { IntrospectedStaticClassFunction } from "./static-function.ts";
+
 import { IntrospectedClassCallback } from "./callback.ts";
 import { IntrospectedProperty, IntrospectedField } from "./property.ts";
 import { IntrospectedNamespace } from "./namespace.ts";
 import {
-    sanitizeIdentifierName,
-    parseTypeIdentifier,
-    resolveTypeIdentifier,
     parseDoc,
     parseMetadata
-} from "../utils/index.ts";
+} from "../utils/gir-parsing.ts";
+
+import { parseTypeIdentifier } from "../utils/types.ts";
+import { resolveTypeIdentifier } from "../utils/type-resolution.ts";
+import { sanitizeIdentifierName } from "../utils/naming.ts";
+
 import { IntrospectedSignal } from "./signal.ts";
 import { FormatGenerator } from "../generators/generator.ts";
 import { GirVisitor } from "../visitor.ts";
