+++ packages/lib/src/index.ts
@@ -9,4 +9,5 @@ export * from './messages.ts'
 export * from './utils/index.ts'
 export * from './gir.ts'
 export * from './gir/index.ts'
-export * from './registry.ts'
+export * from './generators/index.ts'
+export * from './formatters/index.ts'
