+++ packages/lib/src/injections/glib.ts
@@ -1,11 +1,11 @@
 import { IntrospectedNamespace } from "../gir/namespace.ts";
-import {
-    IntrospectedConstructor,
-    IntrospectedFunctionParameter,
-    IntrospectedClassFunction,
-    IntrospectedFunction,
-    IntrospectedDirectAllocationConstructor
-} from "../gir/function.ts";
+
+import { IntrospectedFunctionParameter } from "../gir/parameter.ts";
+import { IntrospectedFunction } from "../gir/function.ts";
+import { IntrospectedClassFunction } from "../gir/class-function.ts";
+import { IntrospectedConstructor } from "../gir/constructor.ts";
+import { IntrospectedDirectAllocationConstructor } from "../gir/direct-allocation-constructor.ts";
+
 import {
     NativeType,
     AnyType,
