+++ packages/lib/src/types/index.ts
@@ -8,6 +8,8 @@ export * from './construct-name.ts'
 export * from './dependency-map.ts'
 export * from './dependency.ts'
 export * from './file-info.ts'
+export * from './format.ts'
+export * from './generator-constructor.ts'
 export * from './gir-alias-element.ts'
 export * from './gir-annotation.ts'
 export * from './gir-any-element.ts'
@@ -44,6 +46,7 @@ export * from './local-name-type.ts'
 export * from './local-name.ts'
 export * from './local-names.ts'
 export * from './metadata.ts'
+export * from './ns-loader.ts'
 export * from './options-base.ts'
 export * from './options-generation.ts'
 export * from './options-load.ts'
