+++ packages/lib/src/types/ns-loader.ts
@@ -1,0 +1,7 @@
+import type { ParsedGir } from "./parsed-gir.ts";
+
+
+export interface NSLoader {
+  load(namespace: string, version: string): ParsedGir | null;
+  loadAll(namespace: string): ParsedGir[];
+}
