+++ packages/lib/src/utils/conflicts.ts
@@ -1,13 +1,13 @@
 import type { IntrospectedNamespace } from "../gir/namespace.ts";
 import type { IntrospectedFunction } from "../gir/function.ts";
-import { IntrospectedClassFunction } from "../gir/function.ts";
-import { IntrospectedConstructor } from "../gir/function.ts";
+import { IntrospectedClassFunction } from "../gir/class-function.ts";
+import { IntrospectedConstructor } from "../gir/constructor.ts";
 import type { TypeIdentifier } from "../gir.ts";
-import { isSubtypeOf } from "./index.ts";
+import { isSubtypeOf } from "./type-resolution.ts";
 import type { IntrospectedBaseClass } from "../gir/introspected-base-class.ts";
-import { IntrospectedStaticClassFunction } from "../gir/function.ts";
-import { IntrospectedVirtualClassFunction } from "../gir/function.ts";
-import { IntrospectedFunctionParameter } from "../gir/function.ts";
+import { IntrospectedStaticClassFunction } from "../gir/static-function.ts";
+import { IntrospectedVirtualClassFunction } from "../gir/virtual-function.ts";
+import { IntrospectedFunctionParameter } from "../gir/parameter.ts";
 import { GirDirection } from "@gi.ts/parser";
 import { ArrayType } from "../gir.ts";
 import { NeverType } from "../gir.ts";
