+++ packages/lib/src/utils/index.ts
@@ -7,6 +7,7 @@ export * from './naming.ts'
 export * from './numbers.ts'
 export * from './objects.ts'
 export * from './path.ts'
+export * from './registry.ts'
 export * from './strings.ts'
 export * from './type-resolution.ts'
 export * from './types.ts'
