+++ packages/lib/src/utils/registry.ts
@@ -1,0 +1,5 @@
+import { SanitizedIdentifiers } from './naming.ts'
+
+export function getSanitizedIdentifiers(): ReadonlyMap<string, string> {
+    return SanitizedIdentifiers
+}
