+++ packages/lib/src/visitor.ts
@@ -1,24 +1,26 @@
-import { TypeExpression } from './gir.ts'
-import { IntrospectedAlias } from './gir/alias.ts'
-import { IntrospectedRecord } from './gir/record.ts'
-import { IntrospectedInterface } from './gir/interface.ts'
-import { IntrospectedClass } from './gir/introspected-class.ts'
-import { IntrospectedBaseClass } from './gir/introspected-base-class.ts'
-import { IntrospectedConstant } from './gir/const.ts'
-import { GirEnumMember, IntrospectedError, IntrospectedEnum } from './gir/enum.ts'
-import {
-    IntrospectedConstructor,
+import type { TypeExpression } from './gir.ts'
+import type { IntrospectedAlias } from './gir/alias.ts'
+import type { IntrospectedRecord } from './gir/record.ts'
+import type { IntrospectedInterface } from './gir/interface.ts'
+import type { IntrospectedClass } from './gir/introspected-class.ts'
+import type { IntrospectedBaseClass } from './gir/introspected-base-class.ts'
+import type { IntrospectedConstant } from './gir/const.ts'
+import type { GirEnumMember, IntrospectedError, IntrospectedEnum } from './gir/enum.ts'
+import type {
     IntrospectedFunctionParameter,
-    IntrospectedFunction,
+} from './gir/parameter.ts'
+import type { IntrospectedStaticClassFunction } from './gir/static-function.ts'
+import type {
     IntrospectedClassFunction,
-    IntrospectedStaticClassFunction,
-    IntrospectedVirtualClassFunction,
-    IntrospectedDirectAllocationConstructor,
-} from './gir/function.ts'
-import { IntrospectedCallback, IntrospectedClassCallback } from './gir/callback.ts'
-import { IntrospectedNamespace } from './gir/namespace.ts'
-import { IntrospectedProperty, IntrospectedField } from './gir/property.ts'
-import { IntrospectedSignal, IntrospectedSignalType } from './gir/signal.ts'
+} from './gir/class-function.ts'
+import type { IntrospectedFunction } from './gir/function.ts'
+import type { IntrospectedConstructor } from './gir/constructor.ts'
+import type { IntrospectedVirtualClassFunction } from './gir/virtual-function.ts'
+import type { IntrospectedDirectAllocationConstructor } from './gir/direct-allocation-constructor.ts'
+import type { IntrospectedCallback, IntrospectedClassCallback } from './gir/callback.ts'
+import type { IntrospectedNamespace } from './gir/namespace.ts'
+import type { IntrospectedProperty, IntrospectedField } from './gir/property.ts'
+import type { IntrospectedSignal, IntrospectedSignalType } from './gir/signal.ts'
 
 export abstract class GirVisitor {
     visitType?: (node: TypeExpression) => TypeExpression
